moving attributes from dedicated folder
