+++ src/View/Components/Dropdown/Items.php
@@ -3,8 +3,8 @@
 namespace TallStackUi\View\Components\Dropdown;
 
 use Illuminate\Contracts\View\View;
+use TallStackUi\Foundation\Attributes\SoftPersonalization;
 use TallStackUi\Foundation\Personalization\Contracts\Personalization;
-use TallStackUi\Foundation\Personalization\SoftPersonalization;
 use TallStackUi\View\Components\BaseComponent;
 
 #[SoftPersonalization('dropdown.items')]
