+++ src/View/Components/Form/Error.php
@@ -3,8 +3,8 @@
 namespace TallStackUi\View\Components\Form;
 
 use Illuminate\Contracts\View\View;
+use TallStackUi\Foundation\Attributes\SoftPersonalization;
 use TallStackUi\Foundation\Personalization\Contracts\Personalization;
-use TallStackUi\Foundation\Personalization\SoftPersonalization;
 use TallStackUi\View\Components\BaseComponent;
 
 #[SoftPersonalization('form.error')]
