+++ tests/Feature/Structure/SoftPersonalizationAttributeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use TallStackUi\Foundation\Personalization\SoftPersonalization;
+use TallStackUi\Foundation\Attributes\SoftPersonalization;
 
 test('can implement the attribute', function (string $index) {
     expect($index)->toHaveAttribute(SoftPersonalization::class);
